requirements.txt
flowlauncher
speedtest-cli